Initialize OptOutStoreSnapshot with zero partitions
This change is to make OptOutStoreSnapshot testable.

If you try to create a new OptOutStoreSnapshot, you will get an
assertion failure when converting the empty heap to a partition.

Until that is fixed, we work around it by starting with 0 partitions.
This means we have to perform a few additional checks on the number of
partitions, but it's not too bad.

Also simplify the code by using a List so we don't have to keep track of
indexes.